+++ CHANGES.md
@@ -10,6 +10,7 @@
 * Files can be uploaded and downloaded in some external apps without needing to download from the Home Access Plus+ app first
 * Users can log out of the app if they want to pass it to another user - [#4](https://github.com/stuajnht/HAP-for-iOS/issues/4)
 * Users are automatically logged back in to the HAP+ server and if they do not use the app for a period of time (or their logon sessions are renewed constantly if the app is in the foreground), so they can continue where they left off almost straight away
+* Users who are using the device during a timetabled lesson are automatically logged out at the end of the lesson. The device needs to be set up in "shared" mode for this to work, and the [timetable plugin for HAP+](https://hap.codeplex.com/wikipage?title=Timetable%20Plugin&referringTitle=Documentation) needs to be installed
 
 ### Changes
 * App restoration now takes place if the app is stopped in the background, so that when opened again the user remains in the same folder location they were before